requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
discord.py==2.6.3
ebaysdk==2.2.0
et_xmlfile==2.0.0
frozenlist==1.7.0
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gspread==6.2.1
httplib2==0.31.0
idna==3.10
lxml==6.0.1
multidict==6.6.4
numpy==2.3.3
oauth2client==4.1.3
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.3.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
xlsxwriter==3.2.9
xlwt==1.3.0
yarl==1.20.1
